Try starting a different test server for every test
